telebot/requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
idna==3.2
lxml==4.6.3
multitasking==0.0.9
numpy==1.21.2
pandas==1.3.2
pycodestyle==2.7.0
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
tornado==6.1
tzlocal==3.0
urllib3==1.26.6
ut==0.2.0
yf==0.0.2
yfinance==0.1.63
